modify: delete group command

+++ src/commands/delete_group_cmd.ts
@@ -1,7 +1,9 @@
+import * as vscode from 'vscode';
 import { command } from '../decorator';
 import { CommandBase } from './CommandBase';
 import { DeleteGroupParams, DeleteGroupResult, IDeleteGroupCmd } from '../types';
 import { getTabGroupDataProvider } from '../provider';
+import { CONFIRM_OPTION } from '../enums';
 
 @command({
   identifier: 'delete.group',
@@ -11,6 +13,18 @@ export class DeleteGroupCmd extends CommandBase<DeleteGroupParams, DeleteGroupRe
 
 async function deleteGroupHandler(params: DeleteGroupParams): Promise<DeleteGroupResult> {
   const tabGroupProvier = getTabGroupDataProvider();
+
+  const group = tabGroupProvier.getGroup(params.id);
+  const confirmResult = await vscode.window.showInformationMessage<{ id: CONFIRM_OPTION; title: string }>(
+    `'${group.toString()}'을 정말 삭제하시겠습니까?`,
+    { id: CONFIRM_OPTION.Ok, title: '예' },
+    { id: CONFIRM_OPTION.Cancel, title: '아니오' }
+  );
+
+  if (confirmResult?.id !== CONFIRM_OPTION.Ok) {
+    return { done: false };
+  }
+
   tabGroupProvier.deleteGroup(params.id);
   tabGroupProvier.refresh();
